finished moving things into snapshot stage
